refactor : 게임 컨트롤러 및 유저 컨트롤러 분리
